Implemented deterministic evaluation for dealer
- Implemented a bitmasked deterministic evaluation method for the
  dealer's hand using the Treys library and lookup tables, replacing the
  previous Monte Carlo simulation approach.
- Fixed a logic error where an equality check (==) was used instead of
  an inclusive inequality (<=) for the 8-card hand limit.
- Validated fix with a 2,000-game trial, achieving a stable 63.15 +-
  2.16% (+- 2 sd) win rate. Although this is almost 10 times slower
  than the previous Monte Carlo method, it achieves about 1% improvement
  in win rate.

+++ backend/app/simulators/perfect_evaluator.py
@@ -80,11 +80,13 @@ def _calculate_base_weight(mset_counts: int, hand_rank_counts: int, rem_rank_cou
     need_counts = mset_counts - hand_rank_counts
     temp_need_counts = need_counts
 
+    # Numba defaults literals to 32-bit. 0xF << 48 (Ace shift) overflows 32-bit ints!
+    # You MUST cast it to 64-bit to prevent corruption on high cards.
     while temp_need_counts:
         lsb = temp_need_counts & -temp_need_counts
         shift = ((custom_bit_length(lsb) - 1) >> 2) << 2
-        weight *= COMB_LUT[rem_rank_counts >> shift & 0xF][(need_counts >> shift) & 0xF]
-        temp_need_counts &= (~(0xF << shift))
+        weight *= COMB_LUT[(rem_rank_counts >> shift) & 0xF][(need_counts >> shift) & 0xF]
+        temp_need_counts &= (~(np.int64(0xF) << np.int64(shift)))
       
     return weight, need_counts
